# gin/docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "howie",
        "title": "Golang Gin API",
        "termsOfService": "https://github.com/hwholiday/test",
        "contact": {},
        "license": {
            "name": "Howie",
            "url": "https://github.com/hwholiday/test"
        },
        "version": "1.0"
    },
    "paths": {
        "/v1/announcement/{id}": {
            "get": {
                "description": "通过ID获取公告信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "announcement"
                ],
                "summary": "获取公告",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Announcement ID",
                        "name": "id",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "announcement"
                ],
                "summary": "删除公告",
                "parameters": [
                    {}
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "model.Announcement": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "create_time": {
                    "type": "integer"
                },
                "expired_time": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                },
                "update_time": {
                    "type": "integer"
                },
                "url": {
                    "type": "string"
                }
            }
        }
    }
}